refactor: ref BurgerConstructor, added useSelector
