+++ src/Components/App/App.jsx
@@ -1,7 +1,7 @@
 import React from "react";
-import { useDispatch, useSelector } from "react-redux";
+// import { useDispatch, useSelector } from "react-redux";
 import { classNames } from "../../helpers/classNames";
-import { getIndredients } from "../../services/actions/getIngredients";
+// import { getIndredients } from "../../services/actions/getIngredients";
 import { ingredientsArr } from "../../utils/prop-types";
 import { BurgerConstructor } from "../BurgerConstructor/BurgerConstructor";
 import { BurgerIngredients } from "../BurgerIngredients/BurgerIngredients";
@@ -10,27 +10,12 @@ import { Header } from "../Header/Header";
 import styles from "./styles.module.css";
 
 function App() {
-  const { data, feedRequest } = useSelector(
-    (state) => state.ingredientsReducer
-  );
-  const dispatch = useDispatch();
-
-  React.useEffect(() => {
-    dispatch(getIndredients());
-  }, [dispatch]);
-
   return (
     <div className={classNames(styles.app, {}, [])}>
       <Header />
       <main className={classNames(styles.main, {}, [])}>
-        {feedRequest ? (
-          <div>Loading...</div>
-        ) : (
-          <>
-            <BurgerIngredients dataList={data} />
-            <BurgerConstructor dataList={data} />
-          </>
-        )}
+        <BurgerIngredients />
+        <BurgerConstructor />
       </main>
       <Footer />
     </div>
