+++ src/Components/BurgerIngredients/BurgerIngredients.jsx
@@ -1,6 +1,8 @@
 import { Tab } from "@ya.praktikum/react-developer-burger-ui-components";
-import React, { createRef, useEffect, useState } from "react";
-import { ingredientsArr } from "../../utils/prop-types";
+import React, { createRef, useState } from "react";
+// import { ingredientsArr } from "../../utils/prop-types";
+import { useDispatch, useSelector } from "react-redux";
+import { getIndredients } from "../../services/actions/getIngredients";
 import { Bread } from "../Bread/Bread";
 import { Filling } from "../Filling/Filling";
 import { Modal } from "../Modal/Modal";
@@ -8,16 +10,18 @@ import { IngredientDetails } from "../ModalIngredients/IngredientDetails";
 import { Sauce } from "../Sauce/Sauce";
 import styles from "./styles.module.css";
 
-export const BurgerIngredients = React.memo(({ dataList }) => {
+export const BurgerIngredients = React.memo(() => {
   const [isVisible, setIsVisible] = useState(false);
   const [modalData, setModalData] = useState({});
   const [current, setCurrent] = useState("one");
-  const [breadList, setBreadList] = useState([]);
-  const [sauceList, setSauceList] = useState([]);
-  const [fillingList, setFillingList] = useState([]);
   const bunRef = createRef(null);
   const sauceRef = createRef(null);
   const fillingRef = createRef(null);
+
+  const { data, feedRequest } = useSelector(
+    (state) => state.ingredientsReducer
+  );
+  const dispatch = useDispatch();
 
   const handleOpenModal = (card) => {
     setModalData(card);
@@ -33,14 +37,9 @@ export const BurgerIngredients = React.memo(({ dataList }) => {
     setCurrent(str);
   };
 
-  useEffect(() => {
-    const bun = dataList.filter((item) => item.type === "bun");
-    const sauce = dataList.filter((item) => item.type === "sauce");
-    const filling = dataList.filter((item) => item.type === "main");
-    setBreadList(bun);
-    setSauceList(sauce);
-    setFillingList(filling);
-  }, [dataList]);
+  React.useEffect(() => {
+    dispatch(getIndredients());
+  }, [dispatch]);
 
   return (
     <section className="pt-10">
@@ -68,71 +67,81 @@ export const BurgerIngredients = React.memo(({ dataList }) => {
           Начинки
         </Tab>
       </div>
-      <div className={styles.container}>
-        <h2 ref={bunRef} className="text text_type_main-medium">
-          Булки
-        </h2>
-        <ul className={styles.grid}>
-          {breadList.map((item) => {
-            return (
-              <Bread
-                key={item._id}
-                name={item.name}
-                price={item.price}
-                image={item.image}
-                imageLarge={item.image_large}
-                calories={item.calories}
-                proteins={item.proteins}
-                fat={item.fat}
-                carbohydrates={item.carbohydrates}
-                onOpen={handleOpenModal}
-              />
-            );
-          })}
-        </ul>
-        <h2 ref={sauceRef} className="mt-10 text text_type_main-medium">
-          Соусы
-        </h2>
-        <ul className={styles.grid}>
-          {sauceList.map((item) => {
-            return (
-              <Sauce
-                key={item._id}
-                name={item.name}
-                price={item.price}
-                image={item.image}
-                imageLarge={item.image_large}
-                calories={item.calories}
-                proteins={item.proteins}
-                fat={item.fat}
-                carbohydrates={item.carbohydrates}
-                onOpen={handleOpenModal}
-              />
-            );
-          })}
-        </ul>
-        <h2 ref={fillingRef} className="mt-10 text text_type_main-medium">
-          Начинка
-        </h2>
-        <ul className={styles.grid}>
-          {fillingList.map((item) => {
-            return (
-              <Filling
-                key={item._id}
-                name={item.name}
-                price={item.price}
-                image={item.image}
-                imageLarge={item.image_large}
-                calories={item.calories}
-                proteins={item.proteins}
-                fat={item.fat}
-                carbohydrates={item.carbohydrates}
-                onOpen={handleOpenModal}
-              />
-            );
-          })}
-        </ul>
-      </div>
+      {feedRequest ? (
+        <div>Loading...</div>
+      ) : (
+        <div className={styles.container}>
+          <h2 ref={bunRef} className="text text_type_main-medium">
+            Булки
+          </h2>
+          <ul className={styles.grid}>
+            {data
+              .filter((item) => item.type === "bun")
+              .map((item) => {
+                return (
+                  <Bread
+                    key={item._id}
+                    name={item.name}
+                    price={item.price}
+                    image={item.image}
+                    imageLarge={item.image_large}
+                    calories={item.calories}
+                    proteins={item.proteins}
+                    fat={item.fat}
+                    carbohydrates={item.carbohydrates}
+                    onOpen={handleOpenModal}
+                  />
+                );
+              })}
+          </ul>
+          <h2 ref={sauceRef} className="mt-10 text text_type_main-medium">
+            Соусы
+          </h2>
+          <ul className={styles.grid}>
+            {data
+              .filter((item) => item.type === "sauce")
+              .map((item) => {
+                return (
+                  <Sauce
+                    key={item._id}
+                    name={item.name}
+                    price={item.price}
+                    image={item.image}
+                    imageLarge={item.image_large}
+                    calories={item.calories}
+                    proteins={item.proteins}
+                    fat={item.fat}
+                    carbohydrates={item.carbohydrates}
+                    onOpen={handleOpenModal}
+                  />
+                );
+              })}
+          </ul>
+          <h2 ref={fillingRef} className="mt-10 text text_type_main-medium">
+            Начинка
+          </h2>
+          <ul className={styles.grid}>
+            {data
+              .filter((item) => item.type === "main")
+              .map((item) => {
+                return (
+                  <Filling
+                    key={item._id}
+                    name={item.name}
+                    price={item.price}
+                    image={item.image}
+                    imageLarge={item.image_large}
+                    calories={item.calories}
+                    proteins={item.proteins}
+                    fat={item.fat}
+                    carbohydrates={item.carbohydrates}
+                    onOpen={handleOpenModal}
+                  />
+                );
+              })}
+          </ul>
+        </div>
+      )}
       <Modal
         children={<IngredientDetails data={modalData} />}
         isOpen={isVisible}
@@ -142,7 +151,3 @@ export const BurgerIngredients = React.memo(({ dataList }) => {
     </section>
   );
 });
-
-BurgerIngredients.propTypes = {
-  dataList: ingredientsArr,
-};
