+++ src/services/store.js
@@ -12,8 +12,6 @@ const composeEnhancers =
     ? window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__({})
     : compose;
 
-// const enhancer = composeEnhancers();
-
 const rootReducer = combineReducers({
   ingredientsReducer,
 });
